+++ attention_sinks/inject_mixin.py
@@ -16,6 +16,7 @@ MODEL_NAME_MAPPING = {
     "gpt_neox": "GPTNeoXModel",
     "gptj": "GPTJModel",
     "mistral": "MistralModel",
+    "qwen": "QWenModel",
 }
 ATTENTION_NAME_MAPPING = {
     "llama": "LlamaAttention",
@@ -24,6 +25,7 @@ ATTENTION_NAME_MAPPING = {
     "gpt_neox": "GPTNeoXAttention",
     "gptj": "GPTJAttention",
     "mistral": "MistralAttention",
+    "qwen": "QWenAttention",
 }
 KV_DIM_MAPPING = {
     "llama": (2, 2),
@@ -32,6 +34,7 @@ KV_DIM_MAPPING = {
     "gpt_neox": (2, 2),
     "gptj": (2, 2),
     "mistral": (2, 2),
+    "qwen": (1, 1),
 }
 
 
@@ -84,6 +87,7 @@ class InjectAttentionSinksMixin:
             gptj_pos_shift_attention_forward,
             llama_pos_shift_attention_forward,
             mistral_pos_shift_attention_forward,
+            qwen_pos_shift_attention_forward,
         )
 
         ATTENTION_FORWARD_MAPPING = {
@@ -93,6 +97,7 @@ class InjectAttentionSinksMixin:
             "gpt_neox": gpt_neox_pos_shift_attention_forward,
             "gptj": gptj_pos_shift_attention_forward,
             "mistral": mistral_pos_shift_attention_forward,
+            "qwen": qwen_pos_shift_attention_forward,
         }
 
         # Not all models require updated attention forwards